backend/requirements.txt
anthropic
python-dotenv
quart==0.17.0
quart-cors==0.4.0
python-telegram-bot==20.7
werkzeug==2.0.3
tqdm
requests
openai>=1.5.0
python-arango>=7.4.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest==8.3.4
pypdf==3.15.0
asyncpg